requirements.txt
streamlit>=1.32.0
anthropic>=0.40.0
sqlglot>=25.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
